# openapi.yaml
openapi: 3.0.3
info:
  title: Logday Server API
  description: |
    API de sync self-hosted para [Logday](https://github.com/jorgebuitragor/logday),
    al estilo Vaultwarden/Bitwarden. Un cliente (desktop, web, móvil, extensión)
    apunta a la URL de una instancia self-hosted y usa esta API para sincronizar
    sus datos entre dispositivos.

    ## Modelo de autenticación

    - No hay auto-registro. Los usuarios los crea un admin (`POST /admin/users`
      o el panel web), salvo el primer admin de la instancia (`/setup` o
      variables de entorno `ADMIN_EMAIL`/`ADMIN_PASSWORD`, ver README).
    - `POST /auth/login` devuelve un `access_token` (JWT de corta duración,
      usado como `Authorization: Bearer <token>` en cada request) y un
      `refresh_token` (opaco, de larga duración, ligado a un `device_id`).
    - Cuando el access token expira, `POST /auth/refresh` cambia el refresh
      token (rotación: el anterior queda invalidado; reusarlo revoca el
      dispositivo — indicio de robo de token).
    - Los TTLs de ambos tokens y el límite de dispositivos por usuario son
      configurables por el admin desde el panel (`internal/settings`), por lo
      que no son constantes fijas — un cliente no debe asumir valores.

    ## Modelo de sync

    Cada recurso (`tasks`, `notes`, `overtime-entries`, `overtime-month-meta`,
    `calendar-events`, `absence-days`, `daily-entries`) expone su propio CRUD
...
    típico:

...
       cursor local.
...

...
    endpoint de contenido en base64.

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /overtime-entries:
    get:
      operationId: listOvertimeEntries
      tags: [Overtime]
      summary: Lista todas las entradas de horas extra del usuario
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OvertimeEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: upsertOvertimeEntryByBody
      tags: [Overtime]
      summary: Crea o actualiza una entrada (upsert por `id` del body)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OvertimeEntryWrite"
      responses:
        "200":
          description: Entrada creada o actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OvertimeEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La entrada pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "409":
          description: Conflicto LWW.
          content:
            text/plain: { schema: { type: string } }
  /overtime-entries/{id}:
    patch:
      operationId: patchOvertimeEntry
      tags: [Overtime]
      summary: Edita campos puntuales de una entrada (LWW por campo)
      description: Mismo protocolo que `PATCH /tasks/{id}` (ver `specs/lww-por-campo`) — solo los campos presentes se evalúan, cada uno por su propio timestamp, nunca responde `409`.
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OvertimeEntryPatch"
      responses:
        "200":
          description: Estado actual de la entrada tras aplicar los campos que ganaron el LWW.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OvertimeEntry"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La entrada pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteOvertimeEntry
      tags: [Overtime]
      summary: Soft-delete de una entrada
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Borrada (tombstone).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: La entrada pertenece a otro usuario.
          content:
            text/plain: { schema: { type: string } }
        "404":
          $ref: "#/components/responses/NotFound"
  /overtime-month-meta:
    get:
      operationId: listOvertimeMonthMeta
      tags: [Overtime]
      summary: Lista los metadatos por mes (colaborador/cédula) del usuario
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OvertimeMonthMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /overtime-month-meta/{yearMonth}:
    patch:
      operationId: patchOvertimeMonthMeta
      tags: [Overtime]
      summary: Crea o edita los metadatos de un mes (LWW por campo)
      description: |
        No hay `id` propio ni `POST` — la clave natural es `(user_id,
        year_month)`, y `yearMonth` en la URL cumple ese rol tanto en la
        escritura como en `/sync/changes`. Por eso este `PATCH` también crea
        el registro si `yearMonth` no existía todavía (mismo comportamiento
        que el `PUT` que reemplaza). Campos presentes se evalúan por su
        propio timestamp (ver `specs/lww-por-campo`); nunca responde `409`.
      parameters:
        - name: yearMonth
          in: path
          required: true
          description: "Formato `YYYY-MM`."
          schema:
            type: string
            example: "2026-08"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OvertimeMonthMetaWrite"
      responses:
        "200":
          description: Estado actual de los metadatos del mes tras aplicar los campos que ganaron el LWW.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OvertimeMonthMeta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteOvertimeMonthMeta
      tags: [Overtime]
      summary: Soft-delete de los metadatos de un mes
      parameters:
        - name: yearMonth
          in: path
          required: true
          schema:
            type: string
            example: "2026-08"
      responses:
        "204":
          description: Borrados (tombstone).
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token devuelto por `/auth/login` o `/auth/refresh`, enviado como `Authorization: Bearer <access_token>`."
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      description: ID generado por el cliente (UUID recomendado) al crear el recurso.
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Body inválido o falla una validación de campo (ver el mensaje).
      content:
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: Falta el header `Authorization`, o el access token es inválido/expiró.
      content:
        text/plain:
          schema:
            type: string
            example: invalid or expired token
...
    NotFound:
      description: El recurso no existe (o ya estaba borrado) para este usuario.
      content:
        text/plain:
          schema:
            type: string
  schemas:
...
    OvertimeEntryWrite:
      type: object
      required: [id, fecha, updated_at]
      properties:
        id:
          type: string
        fecha:
          type: string
          format: date
        solicitada_por:
          type: string
        actividad:
          type: string
        observaciones:
          type: string
        hora_inicio:
          type: string
        hora_final:
          type: string
        total_horas:
          type: number
          format: double
        extras_diurnas:
          type: number
          format: double
        extras_nocturnas:
          type: number
          format: double
        extras_diurnas_festivas:
          type: number
          format: double
        extras_nocturnas_festivas:
          type: number
          format: double
        updated_at:
          type: string
          format: date-time
    OvertimeEntry:
      allOf:
        - $ref: "#/components/schemas/OvertimeEntryWrite"
        - type: object
          properties:
            seq:
              type: integer
              format: int64
            deleted_at:
              type: string
              format: date-time
              nullable: true
    OvertimeEntryPatch:
      type: object
      required: [updated_at]
      description: Todos los campos son opcionales salvo `updated_at` (ver `PATCH /overtime-entries/{id}`).
      properties:
        fecha:
          type: string
          format: date
        solicitada_por:
          type: string
        actividad:
          type: string
        observaciones:
          type: string
        hora_inicio:
          type: string
        hora_final:
          type: string
        total_horas:
          type: number
          format: double
        extras_diurnas:
          type: number
          format: double
        extras_nocturnas:
          type: number
          format: double
        extras_diurnas_festivas:
          type: number
          format: double
        extras_nocturnas_festivas:
          type: number
          format: double
        updated_at:
          type: string
          format: date-time
    OvertimeMonthMetaWrite:
      type: object
      required: [updated_at]
      properties:
        colaborador:
          type: string
        cedula:
          type: string
        updated_at:
          type: string
          format: date-time
    OvertimeMonthMeta:
      allOf:
        - $ref: "#/components/schemas/OvertimeMonthMetaWrite"
        - type: object
          properties:
            year_month:
              type: string
              example: "2026-08"
            seq:
              type: integer
              format: int64
            deleted_at:
              type: string
              format: date-time
              nullable: true

    # --- Calendar ---
...